MyApi/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
bcrypt @ file:///C:/ci/bcrypt_1597936263757/work
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography @ file:///C:/ci/cryptography_1601046905460/work
cssselect==1.1.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
hyperlink @ file:///tmp/build/80754af9/hyperlink_1600360282274/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
incremental==17.5.0
iniconfig @ file:///tmp/build/80754af9/iniconfig_1596827328212/work
itemadapter @ file:///tmp/build/80754af9/itemadapter_1598654773526/work
itemloaders @ file:///tmp/build/80754af9/itemloaders_1599702791960/work
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
lxml @ file:///C:/ci/lxml_1594822774489/work
more-itertools @ file:///tmp/build/80754af9/more-itertools_1598884071673/work
oauthlib==3.1.0
packaging==20.4
parsel==1.6.0
pluggy==0.13.1
Protego @ file:///tmp/build/80754af9/protego_1598657180827/work
py @ file:///tmp/build/80754af9/py_1593446248552/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyDispatcher==2.0.5
PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1594390921726/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pytest==6.0.2
pytest-runner==5.2
python3-openid==3.2.0
pytz==2020.1
pywin32==227
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
Scrapy @ file:///C:/ci/scrapy_1598649882180/work
service-identity @ file:///tmp/build/80754af9/service_identity_1592494375577/work
six==1.15.0
sqlparse==0.3.1
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
Twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
wincertstore==0.2
zope.interface @ file:///C:/ci/zope.interface_1598606196502/work
